Reformatted the macros
There's now a util module which contains half of the modules, so they can be used across the entire crate
